# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Briefcase API
  version: 1.1.0
  description: Organization-scoped file storage, permissions, previews, search, versioning, recovery, and OBO application
    access.
servers:
- url: https://backend.briefcase.teamofsilicons.com/api/v1
  description: The API is served by the backend host. Permanent entry URLs are minted on the application host instead,
    always organization-scoped as https://briefcase.teamofsilicons.com/org/{org_id}/{path}.
security:
- bearerAuth: []
- bearerAuth: []
  testingEnvironmentKey: []
...
paths:
...
  /telemetry:
    post:
      tags: [Telemetry]
      operationId: submitTelemetry
      x-operation-version: 1.0.0
      summary: Submit a bounded operational event
      security: []
      description: Best-effort anonymous relay to Space Station. No table key or IAM credential is required. Events are marked client_reported and never confer authority. At most 100 events per second per process; duplicate IDs are suppressed for 60 seconds. A 204 acknowledges intake, not remote persistence. A disabled or unconfigured collector discards events.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [id, source, operation, stage]
              properties:
                id: {type: string, format: uuid, description: Non-nil event ID retained across retries.}
                source: {type: string, enum: [sdk, cli, daemon, web]}
                operation: {type: string, minLength: 1, maxLength: 96, pattern: '^[A-Za-z0-9_.-]+$', description: Fixed operational name; never credentials or user content. Known credential prefixes are rejected.}
                stage: {type: string, enum: [started, progress, completed, failed, cancelled]}
                testing: {type: boolean, default: false}
                environment_id: {type: [string, 'null'], format: uuid, description: Non-nil public ID; only valid when testing is true.}
                request_id: {type: [string, 'null'], format: uuid}
                duration_ms: {type: [integer, 'null'], minimum: 0}
                status: {type: [integer, 'null'], minimum: 100, maximum: 599}
                attempt: {type: [integer, 'null'], minimum: 0, maximum: 4294967295}
                count: {type: [integer, 'null'], minimum: 0}
                progress: {type: [integer, 'null'], minimum: 0, maximum: 100}
      responses:
        '204': {description: Event accepted for best-effort delivery or collection disabled}
        '429': {description: Process-wide telemetry intake limit reached}
        default:
          $ref: '#/components/responses/Error'
...
  /org/{org_id}/{path}:
    servers:
    - url: https://backend.briefcase.teamofsilicons.com
    get:
      tags:
      - Entries
      operationId: resolvePermanentUrl
      summary: Resolve the clean permanent URL of an entry
      description: The path segment mirrors the folder structure exactly, for example `/org/tos/private/cos:tos/top_secret/this_secret.md`.
        The organization segment must match `X-Org-ID`. Without `disposition` the entry and its effective access
        are returned; with `disposition` the file bytes are streamed. Anything the caller cannot read answers `404`.
      parameters:
      - in: query
        name: test_environment
        description: Public sandbox UUID; unavailable or mismatched selectors never fall back to production.
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OrgId'
      - name: org_id
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: path
        required: true
        schema:
          type: string
          maxLength: 2048
        description: Organization-relative entry path
      - name: disposition
        in: query
        schema:
          type: string
          enum:
          - inline
          - attachment
        description: Omit to receive entry metadata
      - $ref: '#/components/parameters/Range'
      responses:
        '200':
          description: Entry metadata, or complete file bytes when a disposition is requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          $ref: '#/components/responses/PartialFileContent'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.1.0
...
  /public/{org_id}/{path}:
    get:
      operationId: readPublicEntry
      summary: Read or download a publicly shared file or folder
      tags:
      - Permissions
      parameters:
      - in: query
        name: test_environment
        description: Public sandbox UUID; unavailable or mismatched selectors never fall back to production.
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OrgIdPath'
      - in: path
        name: path
        required: true
        schema:
          type: string
      - in: query
        name: view
        schema:
          type: string
          enum:
          - metadata
          - contents
          - inline
          - attachment
          default: metadata
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/OptionalTestingEnvironmentKey'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/PublicEntry'
                - $ref: '#/components/schemas/PublicPage'
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/zstd:
              schema:
                type: string
                format: binary
        default:
          $ref: '#/components/responses/Error'
      security: []
      x-operation-version: 1.1.0
...
components:
  securitySchemes:
    uploadCapability:
      type: apiKey
      in: header
      name: X-Briefcase-Upload-Capability
      description: Narrow private-byte capability issued by an authorized upload reservation
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM access token
    oboAccess:
      type: apiKey
      in: header
      name: X-IAM-OBO-Access-Proof
      description: Short-lived IAM OBO Access proof
    appId:
      type: apiKey
      in: header
      name: X-App-ID
      description: Originating application identifier required together with an OBO Access proof
    testingEnvironmentKey:
      type: apiKey
      in: header
      name: X-Briefcase-App-Secret
      description: 'IAM test Application secret: ask_ followed by 43 URL-safe characters. Selects an already provisioned
        and registered Briefcase testing environment.'
  parameters:
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        type: string
...
    OrgIdPath:
      name: org_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^[a-z0-9_-]{3,50}$
...
    OptionalTestingEnvironmentKey:
      name: X-Briefcase-App-Secret
      in: header
      required: false
      description: Selects a Briefcase test plane; omission selects production
      schema:
        $ref: '#/components/schemas/TestingEnvironmentKeyValue'
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
    Range:
      name: Range
      in: header
      description: Single `bytes=` range. A malformed or multi-range value is ignored and the complete file is returned.
      schema:
        type: string
        pattern: ^bytes=(\d+-\d*|-\d+)$
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PartialFileContent:
      description: Requested byte range of the file
      headers:
        Content-Range:
          schema:
            type: string
          description: bytes {first}-{last}/{complete size}
        Content-Disposition:
          schema:
            type: string
        Accept-Ranges:
          schema:
            const: bytes
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
  schemas:
...
    TestingEnvironmentKeyValue:
      type: string
      pattern: ^ask_[A-Za-z0-9_-]{43}$
      description: IAM testing Application secret selecting the paired Briefcase environment. Never use a production
        secret.
...
    ActorRef:
      type: object
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - carbon
          - silicon
          - application
        id:
          type: string
...
    Entry:
      type: object
      required:
      - id
      - org_id
      - type
      - visibility
      - name
      - path
      - permanent_url
      - effective_access
      description: 'A `traversal` entry is a folder the caller reaches only because something inside it was shared.
        It can be opened and lists what was shared; its owner, timestamps, and effective access are withheld. The
        reserved Public, Private, and tag containers carry no `owner`: they are structure IAM reconciliation maintains
        rather than any member''s property.'
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
        type:
          type: string
          enum:
          - file
          - folder
        visibility:
          type: string
          enum:
          - full
          - traversal
        name:
          type: string
        path:
          type: string
          maxLength: 2048
          description: Organization-relative path without a leading separator
        parent_id:
          type:
          - string
          - 'null'
          format: uuid
        root_type:
          type: string
          enum:
          - public
          - private
          - tag
        tag:
          type:
          - string
          - 'null'
        content_type:
          type:
          - string
          - 'null'
        size:
          type:
          - integer
          - 'null'
          minimum: 0
        render:
          description: Renderer a client should open for a file; `unsupported` means metadata and CRUD only.
          oneOf:
          - type: 'null'
          - type: string
            enum:
            - image
            - video
            - document
            - spreadsheet
            - presentation
            - audio
            - archive
            - code
            - unsupported
        permanent_url:
          type: string
          format: uri
          description: Clean permanent URL that shows the folder structure
        content_url:
          type:
          - string
          - 'null'
          format: uri
          description: Authenticated sandboxed content URL for a file
        download_url:
          type:
          - string
          - 'null'
          format: uri
          description: Authenticated attachment URL for a file
        owner:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ActorRef'
        origin_app_id:
          type:
          - string
          - 'null'
        effective_access:
          type: array
          items:
            type: string
            enum:
            - read
            - write
            - update
            - delete
            - manage_permissions
        created_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type:
          - string
          - 'null'
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    PublicEntry:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - path
      - entry_type
      - content_type
      - size
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        path:
          type: string
        entry_type:
          type: string
          enum:
          - file
          - folder
        content_type:
          type:
          - string
          - 'null'
        size:
          type:
          - integer
          - 'null'
          minimum: 0
    PublicPage:
      type: object
      additionalProperties: false
      required:
      - items
      - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicEntry'
        next_cursor:
          type:
          - string
          - 'null'
...